[Polkadot] [Wallet] add basic save flow of unapproved polkadot tx (#32964)
Resolves brave/brave-browser#51697

This PR implements the functionality required to persist an unsigned Polkadot
send transaction. Note, transactions won't display because of a missing method
implementation in the PolkadotTxMeta::ToTransactionInfo(), which will be added
in a separate PR.

+++ components/brave_wallet/browser/polkadot/polkadot_tx_meta.h
@@ -33,7 +33,7 @@ class PolkadotTxMeta : public TxMeta {
  private:
   std::string recipient_;
   std::string encoded_extrinsic_;
-  base::StrictNumeric<uint64_t> amount_ = uint64_t{0};
+  base::StrictNumeric<uint128_t> amount_ = uint128_t{0};
 };
 
 }  // namespace brave_wallet
